src/requirements/base.txt
Django==2.2
psycopg2-binary
celery==4.3
Pillow
djangorestframework==3.9.4
django-rest-knox==4.0.1
# django-cors-headers
requests